Rename: Trees 페이지 관련 컴포넌트 디렉터리 재구성
