+++ src/components/features/Trees/TreeAddForm/Input/TreeFruitInput/TreeFruitInput.type.ts
@@ -1,0 +1,10 @@
+export type StyledProps = {
+	selected: boolean;
+};
+
+export type Props = {
+	selected: boolean;
+	fruitName: string;
+	imgSrc: string;
+	onChangeSelectedFruit: (fruit: string) => void;
+};
