+++ src/components/features/Trees/TreeAddForm/Input/TreeFruitInput/index.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import TreeFruitInputView from './TreeFruitInputView';
-import { TreeFruitInputModelType } from './type';
+import * as S from './TreeFruitInput.styled';
+import { Props } from './TreeFruitInput.type';
 
-const TreeFruitInput = ({ fruitName, imgSrc, selected, onChangeSelectedFruit }: TreeFruitInputModelType) => {
+const TreeFruitInput = ({ fruitName, imgSrc, selected, onChangeSelectedFruit }: Props) => {
 	const onClick = (event: React.MouseEvent) => {
 		const targetFruit = (event.target as HTMLInputElement).value;
 
@@ -11,7 +11,17 @@ const TreeFruitInput = ({ fruitName, imgSrc, selected, onChangeSelectedFruit }: 
 		}
 	};
 
-	return <TreeFruitInputView fruitName={fruitName} imgSrc={imgSrc} selected={selected} onClick={onClick} />;
+	return (
+		<S.FruitInputContainer selected={selected}>
+			<input type="radio" id={fruitName} name="fruit-radio" value={fruitName} onClick={onClick} />
+			<label htmlFor={fruitName}>
+				<span className="fruit-icon">
+					<img src={imgSrc} alt={`${fruitName} 아이콘`} />
+				</span>
+				<span className="fruit-name">{fruitName}</span>
+			</label>
+		</S.FruitInputContainer>
+	);
 };
 
 export default TreeFruitInput;
