+++ src/components/features/Trees/TreeAddForm/Input/TreeNameInput/TreeNameInput.type.ts
@@ -1,0 +1,6 @@
+export type Props = {
+	placeholder: string;
+	autoFocus?: boolean;
+	treeName: string;
+	onChangeTreeName: (treeName: string) => void;
+};
