+++ src/components/features/Trees/TreeList/TreeItem/TreeItem.type.ts
@@ -1,0 +1,3 @@
+export interface Props {
+	name: string;
+}
