+++ src/components/features/Trees/TreeList/TreeItem/index.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
-import { Link } from 'react-router-dom';
 import * as S from './TreeItem.styled';
+import { Link } from 'react-router-dom';
 import { TreeShapeContainer } from '@/components/features/Trees/TreeAddForm/TreeAddForm.styled';
-import { TreeItemViewType } from './type';
+import { Props } from './TreeItem.type';
 
-const TreeItemView = ({ name }: TreeItemViewType) => {
+const TreeItem = ({ name }: Props) => {
 	return (
 		<S.ItemWrapper>
 			<Link to={name}>
@@ -18,4 +18,4 @@ const TreeItemView = ({ name }: TreeItemViewType) => {
 	);
 };
 
-export default TreeItemView;
+export default TreeItem;
