--- src/components/features/Trees/TreeList/TreeItem/type.ts
@@ -1,3 +1,0 @@
-export interface TreeItemViewType {
-	name: string;
-}
