+++ src/components/features/Trees/TreeList/TreeItemAdd/TreeItemAdd.type.ts
@@ -1,0 +1,3 @@
+export type Props = {
+	onClick: () => void;
+};
