--- src/components/features/Trees/TreeList/TreeItemAdd/type.ts
@@ -1,3 +1,0 @@
-export interface TreeItemAddViewType {
-	onClick: () => void;
-}
