--- src/components/features/Trees/TreeList/type.ts
@@ -1,3 +1,0 @@
-export interface TreeListViewType {
-	handleTreeEditItemClick: () => void;
-}
