--- src/pages/Trees/type.ts
@@ -1,5 +1,0 @@
-export interface TreesType {
-	isAddingTree: boolean;
-	onCloseForm: () => void;
-	onOpenForm: () => void;
-}
